iOS7 compatibility change

+++ RichTextEditor/Source/Categories/UIView+RichTextEditor.m
@@ -33,7 +33,7 @@
 {
     unsigned char pixel[4] = {0};
     CGColorSpaceRef colorSpace = CGColorSpaceCreateDeviceRGB();
-    CGContextRef context = CGBitmapContextCreate(pixel, 1, 1, 8, 4, colorSpace, kCGImageAlphaPremultipliedLast);
+    CGContextRef context = CGBitmapContextCreate(pixel, 1, 1, 8, 4, colorSpace, (uint32_t)kCGImageAlphaPremultipliedLast);
 	
     CGContextTranslateCTM(context, -point.x, -point.y);
